(svn r2046) -Codechange: moved all waypoint code to waypoint.c/waypoint.h -Codechange: rewrote some functions while moving waypoint-stuff -Add: added support for 64k waypoints -Fix: made the waypoint struct a bit more logic (no bit-fucking)

+++ newgrf.c
@@ -31,7 +31,7 @@ GRFFile *_first_grffile;
 int _grffile_count;
 static int _cur_spriteid;
 static int _cur_stage;
-extern int _custom_sprites_base;
+extern uint16 _custom_sprites_base;
 
 static int32 _paramlist[0x7f];
 static int _param_max;
